Change merge_single_qubit_moments_to_phxz to support global phase. (#7405)
This addresses an issue I found while working on #7291 

Currently `merge_single_qubit_moments_to_phxz` will not merge moments
that have a global phase, meaning that moments that could theoretically
be merged are not. I updated it so that global phase is also supported.
I added two tests to show the issue.

Perhaps this situation is not common right now, but if #7383 is
committed and the flag is used, it may become more common. @daxfohl

Note that I'm not merging the global phase operations in a single one,
because I'm thinking some phases could be parametrized; should I attempt
to do this?