Fixing MS gate inverse is not implemented correctly (Issue 7432) (#7466)
Fixing MS gate inverse is not implemented correctly (Issue
https://github.com/quantumlib/Cirq/issues/7432)

The MS gate has a theta entanglement angle parameter in it's signature,
with a 0.25 default value. The problems that need to be corrected here
is that when implementing the inverse MS gate the theta parameter is not
applied, and instead implicitly its default value is being used when
specifying inverse of MS in terms of MS.

Code and tests were updated.

---------

Co-authored-by: Michael Hucka <mhucka@caltech.edu>